+++ features/keychain/module/__tests__/binauth.test.js
@@ -1,11 +1,14 @@
 import { mnemonicToSeed } from 'bip39'
 
-import { EXODUS_KEY_IDS } from '../key-identifier'
+import { EXODUS_KEY_IDS } from '@exodus/key-ids'
 import createKeychain from './create-keychain'
+import { getSeedId } from '../crypto/seed-id'
 
 const seed = mnemonicToSeed(
   'menu memory fury language physical wonder dog valid smart edge decrease worth'
 )
+
+const seedId = getSeedId(seed)
 
 describe('binauth', () => {
   it('should create binauth key', async () => {
@@ -15,7 +18,7 @@ describe('binauth', () => {
     const sodiumEncryptor = await keychain.createSodiumEncryptor(EXODUS_KEY_IDS.TELEMETRY)
     const {
       sign: { publicKey },
-    } = await sodiumEncryptor.getSodiumKeysFromSeed()
+    } = await sodiumEncryptor.getSodiumKeysFromSeed({ seedId })
     const expectedPublicKey = Buffer.from(
       'eeab6c9e861ed9f3a7f7917f6d972032e3e4d7a433eb6bc30f4b488ee13682c7',
       'hex'
@@ -25,6 +28,7 @@ describe('binauth', () => {
     // Client signing the challenge
     const challenge = Buffer.from('aabbccc', 'hex')
     const signedChallenge = await sodiumEncryptor.sign({
+      seedId,
       data: challenge,
     })
     const expectedSignedChallenge = Buffer.from(
@@ -41,7 +45,7 @@ describe('binauth', () => {
     const keyId = EXODUS_KEY_IDS.TELEMETRY
     const {
       sign: { publicKey },
-    } = await keychain.sodium.getSodiumKeysFromSeed({ keyId })
+    } = await keychain.sodium.getSodiumKeysFromSeed({ seedId, keyId })
 
     const expectedPublicKey = Buffer.from(
       'eeab6c9e861ed9f3a7f7917f6d972032e3e4d7a433eb6bc30f4b488ee13682c7',
@@ -51,7 +55,11 @@ describe('binauth', () => {
 
     // Client signing the challenge
     const challenge = Buffer.from('aabbccc', 'hex')
-    const signedChallenge = await keychain.sodium.sign({ keyId, data: challenge })
+    const signedChallenge = await keychain.sodium.sign({
+      seedId,
+      keyId,
+      data: challenge,
+    })
 
     const expectedSignedChallenge = Buffer.from(
       'f87037abf6dd8e46cc691880c008ffa5646ba8bf9f523339a503e16b8f6c92c647e00940804ae64770456e8211c18e27234371e9a5f62505f6f50feafcbb2d0faabbcc',