+++ libraries/key-identifier/src/key-identifier.js
@@ -11,7 +11,7 @@ import { assertValidDerivationPath } from '@exodus/key-utils'
 const SUPPORTED_KDFS = new Set(['BIP32', 'SLIP10'])
 const SUPPORTED_KEY_TYPES = new Set(['legacy', 'nacl', 'secp256k1'])
 
-export class KeyIdentifier {
+export default class KeyIdentifier {
   constructor({ derivationAlgorithm, derivationPath, assetName, keyType }) {
     assert(typeof derivationAlgorithm === 'string', 'derivationAlgorithm not a string')
     assert(
@@ -66,48 +66,4 @@ export class KeyIdentifier {
       (fieldName) => keyIdA[fieldName] !== keyIdB[fieldName]
     )
   }
-}
-
-const EXO = Number.parseInt(Buffer.from('exo').toString('hex'), '16')
-export const EXODUS_KEY_IDS = Object.freeze({
-  WALLET_INFO: new KeyIdentifier({
-    derivationAlgorithm: 'BIP32',
-    derivationPath: `m/${EXO}'/1'/0`,
-    keyType: 'nacl',
-  }),
-  BACKUP_FILE: new KeyIdentifier({
-    derivationAlgorithm: 'BIP32',
-    derivationPath: `m/${EXO}'/1'/1`,
-    keyType: 'nacl',
-  }),
-  FUSION: new KeyIdentifier({
-    derivationAlgorithm: 'SLIP10',
-    derivationPath: `m/${EXO}'/2'/0'`,
-    keyType: 'nacl',
-  }),
-  '2FA_MODE': new KeyIdentifier({
-    derivationAlgorithm: 'SLIP10',
-    derivationPath: `m/${EXO}'/2'/1'`,
-    keyType: 'nacl',
-  }),
-  TELEMETRY: new KeyIdentifier({
-    derivationAlgorithm: 'SLIP10',
-    derivationPath: `m/${EXO}'/2'/3'`,
-    keyType: 'nacl',
-  }),
-  SEEDLESS: new KeyIdentifier({
-    derivationAlgorithm: 'SLIP10',
-    derivationPath: `m/${EXO}'/5'/0'`,
-    keyType: 'nacl',
-  }),
-})
-
-export const createKeyIdentifierForExodus = ({ exoType }) => {
-  assert(typeof exoType === 'string', 'exotype must be of type string')
-  const keyId = EXODUS_KEY_IDS[exoType]
-  if (!keyId) {
-    throw new TypeError('Invalid exodus key requested')
-  }
-
-  return keyId
 }