+++ features/keychain/CHANGELOG.md
@@ -1,7 +1,20 @@
+# 6.1.0 (2024-02-20)
+
+### Features
+
+- keychain api, to replace the one from the wallet feature ([#64](https://github.com/ExodusMovement/exodus-oss/issues/64)) ([f5fd3e1cc](https://github.com/ExodusMovement/exodus-oss/commit/f5fd3e1cc5d813c6da527ab972b5402c0a2508e3))
+- add separate key-identifier package ([#61](https://github.com/ExodusMovement/exodus-oss/issues/61)) ([49dc7f6](https://github.com/ExodusMovement/exodus-oss/commit/49dc7f66040a9ecaba1d971858764532d695bbdc))
+
 # Change Log
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [6.0.0](https://github.com/ExodusMovement/exodus-oss/compare/@exodus/keychain@5.0.1...@exodus/keychain@6.0.0) (2024-02-15)
+
+### ⚠ BREAKING CHANGES
+
+- add multi-seed support ([#50](https://github.com/ExodusMovement/exodus-oss/issues/50)) ([82e5cf2e](https://github.com/ExodusMovement/exodus-oss/commit/82e5cf2e1b9334af150bf2a8bba4abbc0f74c764))
 
 ## [5.0.1](https://github.com/ExodusMovement/exodus-oss/compare/@exodus/keychain@5.0.0...@exodus/keychain@5.0.1) (2024-01-26)